+++ exonum/src/runtime/types.rs
@@ -19,8 +19,7 @@ use std::{borrow::Cow, fmt::Display, str::FromStr};
 use crate::{
     blockchain::ConsensusConfig,
     helpers::ValidateInput,
-    impl_binary_key_for_binary_value,
-    merkledb::{is_allowed_latin1_char, is_valid_index_name, BinaryValue},
+    merkledb::{is_allowed_latin1_char, is_valid_index_name, BinaryValue, impl_binary_key_for_binary_value},
     proto::schema,
 };
