+++ exonum/src/sandbox/mod.rs
@@ -14,6 +14,7 @@
 
 use bit_vec::BitVec;
 use exonum_merkledb::{BinaryValue, HashTag, IndexAccess, MapProof, ObjectHash, TemporaryDB};
+use exonum_proto::impl_binary_value_for_pb_message;
 use futures::{sync::mpsc, Async, Future, Sink, Stream};
 
 use std::{
